refactor: 로딩 컴포넌트 지우고 스타일드컴포넌트에 추가
